Bug 1858496 - Crash on OOM in ICU; a=diannaS
We're seeing inconsistent handling of OOMs in the ICU library. This
patch changes the behaviour to crash on OOM rather than allowing
ICU to handle the allocation failure. The inconsistent handling in ICU
could lead to ICU being in an inconsistent state which could in turn
cause security problems. The safer alternative is to crash, but it's
possible this will lead to too high of crash rate.

Original Revision: https://phabricator.services.mozilla.com/D191892

Differential Revision: https://phabricator.services.mozilla.com/D202498