Bug 1896944 - Restore $TmpVal after using it for a different purpose …
…r=nalexander! a=pascalc

This is a bit ugly. The better fix for this issue is https://phabricator.services.mozilla.com/D211452 but that fix is a bit more invasive whereas this fix is a bit safer for uplift. So this patch will be uplifted and the other one will clean things up but won't be uplifted.

Differential Revision: https://phabricator.services.mozilla.com/D211608