Avoid trying to install dist toolchains when bootstrapping from a spidermonkey standalone tree.  a=pascalc The auto-bootstrap code gracefully handles the lack of information as to what toolchains are available in spidermonkey standalone trees, but the fallback code path for when toolchains are not to be unpacked doesn't handle the situation.
By making the toolchains themselves define whether they are to be
unpacked, we can remove the fallback code path and use auto-bootstrap in
all cases.

This makes `mach artifact toolchain --from-build` stop requiring a
manual `--no-unpack`, but `--from-task` still needs it because it
can't infer that information, as it doesn't have the task graph.