Bug 1918096 - Integrate SCContentSharingPicker into ScreenCapturerSck.  a=RyanVM This works by creating ScreenCapturerSck with a sck_allow_system_picker DesktopCaptureOptions flag.
The flow is similar to the Pipewire capturer in that SelectSource has no effect.
Start() brings up the system picker, and:
- automatically starts capture on selection, also reconfigures on future
  selection changes.
- signals a stop by setting the permanent error flag on cancel or error.

The system picker is configured for both single display and multiple windows to
give users as much power over what they share as possible. There is an
application mode also, but other browsers are not allowing access to that as of
now.