Bug 1918096 - In ScreenCapturerSck adapt to sources that change resolution.  a=RyanVM If the resolution is smaller than the allocated surface, we crop. If the resolution is larger than the allocated surface, we reconfigure with a larger surface.
The allocated surface has the size we told it to have in the
SCStreamConfiguration, in pixels.

If the source is a screen, the size is pretty static. Changing display settings
  could affect it.
If the source is a window, the size is initially the size of the window. The
  user resizing the window affects the size.
If the source is multiple windows, the size initially seems to be that of the
  display they're sitting on. Windows across multiple displays are not captured
  into the same surface, as of macOS 14.