support tab group drag/drop in list all tabs menu r=dao,dwalker,tabbrowser-reviewers,extension-reviewers,robwu The "list all tabs" menu displayed tab groups and allowed drag/dropping tabs around, but an error would be thrown when trying to drag an entire tab group. This patch allows dragging tab groups in the "list all tabs" menu and dropping them elsewhere.
This accounts for a number of edge cases:
- drag tab group to the end of the menu
- drag tab right before a tab group
- drag tab right after a collapsed tab group
- drag an ungrouped tab into the position of the first tab of a tab group

The tab group menu items didn't have a `tab` property like the tab menu items. That property is what gets passed to the drag-drop code to do its work. I made `tab` and `tabGroup` properties available on the `toolbaritem` elements and included some helper functions to access them consistently.

With that in place, tab groups would generally drag and drop OK, but the "list all tabs" menu wouldn't update. The menu needs to be rebuilt when a tab group moves, so I started listening for the `TabGroupMoved` event.

TabsPanel._onDrop was using `_tPos` on the menu item's `tab`, but that isn't present on tab group labels. I tried modifying the logic to use the `elementIndex`, but that led to edge cases with `Tabbrowser.moveTabTo`. Since the TabsPanel was already tracking the drop direction, I switched to using moveTabBefore/moveTabAfter for simplicity.

The TabsListBase._addTab method was still assuming that the tab strip was a linear list of tabs, so it wasn't doing great at handling the scenario of dropping a tab next to a tab group label. The drop could result in 1) no change, 2) tab moves correctly but it becomes the last item of the "list all tabs" menu, 3) console error. I switched to using the <tabs> element's `findNextTab`. I found an edge case where Tabbrowser.moveTabAfter would move a tab into the first position of a collapsed tab group instead of moving the tab after the whole tab group. I updated Tabbrowser.#moveTabNextTo for that case.

I encountered an edge case when dragging a tab in "list all tabs" between two pinned tabs when the first item in the unpinned tab strip is a tab group. The dropped tab was being inserted into the group as the first grouped tab instead of dropping as an ungrouped tab. I added a general fix in Tabbrowser.#moveTabNextTo -- not sure if other code paths would be affected by this bug or just the "list all tabs" menu

When we added tab groups to the "list all tabs" menu in bug 1908431, this had two side effects:
1. the hidden tabs menus would show tab group labels if a hidden tab was in an expanded tab group
2. hidden tabs would NOT show up in the hidden tabs menus if they were in a collapsed tab group

Our team decided that (1) was desirable because it added helpful context to those hidden tabs. (2) is not desirable, so this patch tries to ensure that hidden tabs always appear in the different hidden tabs lists even if they are also in collapsed tab groups.

For consistency with tabs in the "list all tabs" menu and consistency with the tab strip, middle-clicking on a tab group label in the "list all tabs" menu will now save and close the tab group.